Azure_AI_Services/requirements.txt
flask==3.1.0
azure-ai-vision-imageanalysis
azure-core
azure-ai-documentintelligence
azure-ai-textanalytics
azure-ai-translation-text
azure-cognitiveservices-speech
